Adding tree functionality, attack, arm, planets
Adding the ability to attack and mine trees. Adding made planets and assets. Adding animations.

+++ Andromeda/Assets/Scripts/GameManager.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public class GameManager : MonoBehaviour
 {
-    public int health = 100;
+    public float health = 100;
 }
